+++ spec/javascripts/components/projects/side_bar/side_bar_spec.js
@@ -1,0 +1,29 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import { SideBar } from 'components/projects/SideBar/index';
+
+describe('<SideBar />', () => {
+  const render = props => {
+    const defaultProps = {
+      reverseColumns: sinon.stub()
+    };
+
+    return shallow(<SideBar {...defaultProps} {...props } />);
+  };
+
+  it('renders the component', () => {
+    const wrapper = render();
+
+    expect(wrapper).toExist();
+  });
+
+  describe('when reverse button is clicked', () => {
+    it('call reverseColumns', () => {
+      const reverseColumns = sinon.stub();
+      const wrapper = render({ reverseColumns });
+
+      wrapper.find('[data-id="reverse-button"]').simulate('click');
+      expect(reverseColumns).toHaveBeenCalled();
+    });
+  });
+});